requirements.txt
feedparser>=6.0.0
praw>=7.7.0
requests>=2.31.0
python-dotenv>=1.0.0
google-genai>=1.0.0
rich>=13.0.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
